LayoutTests/fast/parser/entity-end-iframe-tag.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../js/resources/js-test-pre.js"></script>
</head>
<body>
<iframe id="myiframe">&lt;/iframe></iframe>
<p id="description"></p>
<div id="console"></div>
<script>
description("Test parsing of entity-escaped &lt;/iframe&gt; tag for <a href=\"https://bugs.webkit.org/show_bug.cgi?id=7931\">Bug 7931: Escaped elements within a textarea block can cause the textarea box to be closed prematurely</a>");

var expectedResult = '<' + '/iframe>';
shouldBe('document.getElementById("myiframe").textContent', 'expectedResult');
</script>
<script src="../js/resources/js-test-post.js"></script>
</body>
</html>
