third_party/blink/web_tests/fast/parser/entity-end-iframe-tag-expected.txt

Test parsing of entity-escaped </iframe> tag for Bug 7931: Escaped elements within a textarea block can cause the textarea box to be closed prematurely

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL document.getElementById("myiframe").textContent should be </iframe>. Was &lt;/iframe>.
PASS successfullyParsed is true

TEST COMPLETE
